Completed wrapper around the Matelso MRS API Version 2.0.
The WSDL on which this was based can be found here:

  https://www.matelso.de/Partnerbereich/Matelso_MRS_v_2_0.asmx?WSDL

It provides all that is necessary to create subscribers, assign
vanity numbers to subscribers and route vanity numbers to a
destination number. In addition, subscribers can be viewed although
the association to vanity numbers can not be obtained using this
interface.

Also deletion of vanity numbers and subscribers is also possible.

Handling callbacks from Matelso is not covered by this gem. Since
this requires a way of managing subscribers and knowing which
vanity numbers they have.